Remove duplicated GKE blueprint; use pre_8_0
The master branch of the cloudbolt-forge repo is now for use by CloudBolt 8.0 instances, including
the public content library that ships out of the box in 8.0.

The pre-8.0 version of this blueprint is available in the pre_8_0 branch. This code is used by the
content library provided to versions of CloudBolt older than 8.0.

[#156581283]