Properly disable selinux in the CB BP
Change the path of the SE linux file to modify from /etc/sysconfig/selinux to /etc/selinux/config. In some cases,  both exist as real, non-symlink files, and /etc/selinux/config is the right one.

Also, do not reboot after disabling selinux in the config file, just change the runtime configuration (with `setenforce 0`). Rebooting from a remote scrpit causes the next step in the BP to fail. If we need to reboot for some reason, we can add the "Reboot server" CB Plug-in to the BP.