Only show configured environments as options for the resource groups per resource action
When a user is using the auto-create resource group per resource action in a blueprint, they only care about selecting an environment that is available/configured for one of the PSSIs in this blueprint. So, only show those as options, as opposed to the previous logic that would show all possible azure envs.

[#163275713]