Add Blueprint to deploy multi-node Kubernetes clusters
The Plug-in in this Blueprint uses RKE to confiure any number of provisioned servers as nodes in a Kubnernetes cluster. The Plug-in assigns the controlplane, etcd and worker roles appropriately, given how many nodes are in the cluster. Corresponding CloudBolt objects are also created so that the cluster can be managed from CloudBolt.

[DEV-13553, DEV-13843]