UI Extension for managing Scheduled Snapshots
This UI Extension includes a Server tab for configuring
the details of scheduled snapshots for each supported Server
(those whose RH can manage snapshots), namely a schedule and
maximum number of snapshots, for users that have permission to
manage snapshots. It also ensures the existence of a Recurring
Job that is used to actually take those snapshots, by running
every hour and looking for Servers scheduled for that hour. The
Recurring Job will also delete old snapshots if the configured
maximum # of snapshots (may default to the Recurring Job's
value) has been exceeded for any Server.

This is intended to replace the usual snapshot functionality
in the product when installed, althought that will only happen
when it is downloaded onto the current version, since that
requires product code changes.

[DEV-16098]